devices: Move htm handling out of resolve_board
We had a layering violation between the resolve_board and file_parser
modules in the devices package. Knowledge of how to decode and find htm
files was leaked into the resolve_board module. This commit moves all
the functions to do with handling the htm file into the file_parser
module.
